pyproject.toml
[tool.poetry]
name = "palisade_he_cnn"
version = "0.1.0"
description = ""
authors = [
    "Vikram Saraph <vikram.saraph@jhuapl.edu>",
    "Vivian Maloney <vivian.maloney@jhuapl.edu>",
    "Freddy Obrecht <freddy.obrecht@jhuapl.edu>",
    "Kate Tallaksen <kate.tallaksen@jhuapl.edu>",
    "Prathibha Rama <prathibha.rama@jhuapl.edu>"
]
readme = "README.md"
packages = [
    {include = "palisade_he_cnn"}
]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^1.13.1"
torchvision = "^0.14.1"
optuna = "^3.2.0"
joblib = "^1.3.2"
pytest = "^7.4.0"
openfhe = {path = "../palisade-python/wheelhouse/OpenFHE-1.0.5a0-cp310-cp310-manylinux_2_17_x86_64.manylinux2014_x86_64.whl"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
